+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to Pi Dock are documented here.
+All notable changes to Pi Coding Agent are documented here.
 
 ## [0.2.4]
 
@@ -19,7 +19,7 @@ All notable changes to Pi Dock are documented here.
 ### Added
 
 - Pi terminals now export `EDITOR` and `VISUAL` so Pi's external-editor flow opens in VS Code
-- New `piDock.editorCommand` setting to configure the editor command
+- New `piAgent.editorCommand` setting to configure the editor command
 
 ## [0.2.1]